+++ lib/src/_columns.dart
@@ -1,28 +1,14 @@
 import 'package:domino/domino.dart';
 
-import '_base.dart';
+import '_modifiers.dart';
 
 // https://bulma.io/documentation/columns/basics/
 
-class Columns extends BulmaComponent {
-  final List<Column> _columns;
-  Columns(List<Column> columns) : _columns = columns;
+Element columns({Mods mods, content}) =>
+    apply(new Element('div', classes: ['columns'], content: content), mods);
 
-  @override
-  Element createElement(BuildContext context) {
-    return new Element('div', classes: ['columns'], content: _columns);
-  }
-}
-
-class Column extends BulmaComponent {
-  final _content;
-  Column({dynamic content}) : _content = content;
-
-  @override
-  Element createElement(BuildContext context) {
-    return new Element('div', classes: ['column'], content: _content);
-  }
-}
+Element column({Mods mods, content}) =>
+    apply(new Element('div', classes: ['column'], content: content), mods);
 
 // TODO: add Basics
 // TODO: add Sizes