+++ lib/src/_elements.dart
@@ -1,67 +1,25 @@
 import 'package:domino/domino.dart';
 
-import '_base.dart';
 import '_modifiers.dart';
 
 // https://bulma.io/documentation/elements/box/
 
-class Box extends BulmaComponent {
-  final _content;
+Element box({Mods mods, content}) =>
+    apply(new Element('div', classes: ['box'], content: content), mods);
 
-  Box({content}) : _content = content;
+Element button({
+  Mods mods,
+  EventHandler onClick,
+  dynamic content,
+}) =>
+    apply(new Element('button', classes: ['button'], content: content), mods)
+      ..onClick(onClick);
 
-  @override
-  Element createElement(BuildContext context) {
-    return new Element('div', classes: ['box'], content: _content);
-  }
-}
+Element content({Mods mods, content}) =>
+    apply(new Element('content', content: content), mods);
 
-class Button extends BulmaComponent {
-  final EventHandler _onClick;
-  final _content;
-
-  Button({
-    EventHandler onClick,
-    dynamic content,
-    TypeMod typeMod,
-  })
-      : _onClick = onClick,
-        _content = content {
-    this.typeMod = typeMod;
-  }
-
-  @override
-  Element createElement(BuildContext context) {
-    return new Element(
-      'button',
-      classes: ['button'],
-      content: _content,
-    )..onClick(_onClick);
-  }
-}
-
-class Content extends BulmaComponent {
-  final _content;
-
-  Content({content}) : _content = content;
-
-  @override
-  Element createElement(BuildContext context) {
-    return new Element('content', content: _content);
-  }
-}
-
-class Delete extends BulmaComponent {
-  final String _tag;
-
-  Delete() : _tag = 'a';
-  Delete.button() : _tag = 'button';
-
-  @override
-  Element createElement(BuildContext context) {
-    return new Element(_tag, classes: ['delete']);
-  }
-}
+Element delete({Mods mods}) =>
+    apply(new Element('button', classes: ['delete']), mods);
 
 // TODO: add icon
 // TODO: add image