+++ CHANGELOG.md
@@ -1,5 +1,12 @@
 # Changelog
 
+## 0.2.0
+
+**Breaking Changes:**
+
+- Removed `BulmaComponent`, clients should use `Mods` instead.
+- Removed most `Component` classes, using simple methods instead in simple cases.
+
 ## 0.1.0
 
 - `Input.value` to provide the current value in the `<input>` field.